+++ R/output-csv.R
@@ -77,7 +77,7 @@ lazyframe__sink_csv <- function(
   quote_style = c("necessary", "always", "never", "non_numeric"),
   maintain_order = TRUE,
   storage_options = NULL,
-  retries = 2,
+  retries = deprecated(),
   sync_on_close = c("none", "data", "all"),
   mkdir = FALSE,
   engine = c("auto", "in-memory", "streaming"),
@@ -150,12 +150,30 @@ lazyframe__lazy_sink_csv <- function(
   quote_style = c("necessary", "always", "never", "non_numeric"),
   maintain_order = TRUE,
   storage_options = NULL,
-  retries = 2,
+  retries = deprecated(),
   sync_on_close = c("none", "data", "all"),
   mkdir = FALSE
 ) {
   wrap({
     check_dots_empty0(...)
+
+    if (is_present(retries)) {
+      deprecate_warn(
+        c(
+          `!` = sprintf(
+            "The %s argument is deprecated as of %s 1.9.0.",
+            format_arg("retries"),
+            format_pkg("polars")
+          ),
+          i = sprintf(
+            "Specify %s in %s instead.",
+            format_code("max_retries"),
+            format_arg("storage_options")
+          )
+        )
+      )
+      storage_options <- c(storage_options, max_retries = as.character(retries))
+    }
 
     target <- arg_to_sink_target(path)
     check_arg_is_1byte("separator", separator)
@@ -188,8 +206,7 @@ lazyframe__lazy_sink_csv <- function(
       maintain_order = maintain_order,
       sync_on_close = sync_on_close,
       mkdir = mkdir,
-      storage_options = storage_options,
-      retries = retries
+      storage_options = storage_options
     )
   })
 }
@@ -225,7 +242,7 @@ dataframe__write_csv <- function(
   null_value = "",
   quote_style = c("necessary", "always", "never", "non_numeric"),
   storage_options = NULL,
-  retries = 2
+  retries = deprecated()
 ) {
   wrap({
     check_dots_empty0(...)