+++ R/output-ipc.R
@@ -34,7 +34,7 @@ lazyframe__sink_ipc <- function(
   compat_level = c("newest", "oldest"),
   maintain_order = TRUE,
   storage_options = NULL,
-  retries = 2,
+  retries = deprecated(),
   sync_on_close = c("none", "data", "all"),
   mkdir = FALSE,
   engine = c("auto", "in-memory", "streaming"),
@@ -88,12 +88,30 @@ lazyframe__lazy_sink_ipc <- function(
   compat_level = c("newest", "oldest"),
   maintain_order = TRUE,
   storage_options = NULL,
-  retries = 2,
+  retries = deprecated(),
   sync_on_close = c("none", "data", "all"),
   mkdir = FALSE
 ) {
   wrap({
     check_dots_empty0(...)
+
+    if (is_present(retries)) {
+      deprecate_warn(
+        c(
+          `!` = sprintf(
+            "The %s argument is deprecated as of %s 1.9.0.",
+            format_arg("retries"),
+            format_pkg("polars")
+          ),
+          i = sprintf(
+            "Specify %s in %s instead.",
+            format_code("max_retries"),
+            format_arg("storage_options")
+          )
+        )
+      )
+      storage_options <- c(storage_options, max_retries = as.character(retries))
+    }
 
     compat_level <- use_option_if_missing(
       compat_level,
@@ -119,8 +137,7 @@ lazyframe__lazy_sink_ipc <- function(
       maintain_order = maintain_order,
       sync_on_close = sync_on_close,
       mkdir = mkdir,
-      storage_options = storage_options,
-      retries = retries
+      storage_options = storage_options
     )
   })
 }
@@ -141,7 +158,7 @@ dataframe__write_ipc <- function(
   compression = c("zstd", "lz4", "uncompressed"),
   compat_level = c("newest", "oldest"),
   storage_options = NULL,
-  retries = 2
+  retries = deprecated()
 ) {
   wrap({
     check_dots_empty0(...)