+++ R/output-parquet.R
@@ -84,7 +84,7 @@ lazyframe__sink_parquet <- function(
   data_page_size = NULL,
   maintain_order = TRUE,
   storage_options = NULL,
-  retries = 2,
+  retries = deprecated(),
   sync_on_close = c("none", "data", "all"),
   mkdir = FALSE,
   engine = c("auto", "in-memory", "streaming"),
@@ -138,12 +138,30 @@ lazyframe__lazy_sink_parquet <- function(
   data_page_size = NULL,
   maintain_order = TRUE,
   storage_options = NULL,
-  retries = 2,
+  retries = deprecated(),
   sync_on_close = c("none", "data", "all"),
   mkdir = FALSE
 ) {
   wrap({
     check_dots_empty0(...)
+
+    if (is_present(retries)) {
+      deprecate_warn(
+        c(
+          `!` = sprintf(
+            "The %s argument is deprecated as of %s 1.9.0.",
+            format_arg("retries"),
+            format_pkg("polars")
+          ),
+          i = sprintf(
+            "Specify %s in %s instead.",
+            format_code("max_retries"),
+            format_arg("storage_options")
+          )
+        )
+      )
+      storage_options <- c(storage_options, max_retries = as.character(retries))
+    }
 
     target <- arg_to_sink_target(path)
     compression <- arg_match0(
@@ -187,8 +205,7 @@ lazyframe__lazy_sink_parquet <- function(
       maintain_order = maintain_order,
       sync_on_close = sync_on_close,
       mkdir = mkdir,
-      storage_options = storage_options,
-      retries = retries
+      storage_options = storage_options
     )
   })
 }
@@ -230,7 +247,7 @@ dataframe__write_parquet <- function(
   partition_by = NULL,
   partition_chunk_size_bytes = 4294967296,
   storage_options = NULL,
-  retries = 2,
+  retries = deprecated(),
   mkdir = FALSE
 ) {
   wrap({